+++ src/tmp.cpp
@@ -1,3 +1,3 @@
-#include "NewName/tmp.hpp"
+#include "project/tmp.hpp"
 
 int tmp::add(int a, int b) { return a + b; }
